+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## v1.4.1
 
+### Added
+
+-   Total swing count
+
 ### Fixed
 
+-   Missing subname display
 -   Outside object not being detected
 
 ## v1.4.0 [2021-03-04]